Garlic: fix static_cast type
Use std::ptrdiff_t instead of int

> std::ptrdiff_t is used for pointer arithmetic and array indexing,
> if negative values are possible. Programs that use other types,
> such as int, may fail on, e.g. 64-bit systems when the index exceeds INT_MAX
> or if it relies on 32-bit modular arithmetic.
http://en.cppreference.com/w/cpp/types/ptrdiff_t

+++ src/core/router/garlic.cc
@@ -615,7 +615,7 @@ void GarlicDestination::HandleGarlicPayload(
     LOG(debug) << "GarlicDestination: garlic payload offset=" << offset;
 
     // Sanity test
-    if ((buf + offset) - buf1 > static_cast<int>(len))
+    if ((buf + offset) - buf1 > static_cast<std::ptrdiff_t>(len))
       // TODO(anonimal): in this case, do we want to throw out the entire payload? If not, don't throw.
       throw std::runtime_error("GarlicDestination: clove is too long");
